Created separate class, AdminController, and moved method to show all members from UserController.
